chat feedback hook
