+++ web/src/app/chat/components/ChatPage.tsx
@@ -19,7 +19,6 @@ import { usePopup } from "@/components/admin/connectors/Popup";
 import { SEARCH_PARAM_NAMES } from "@/app/chat/services/searchParams";
 import { useFederatedConnectors, useFilters, useLlmManager } from "@/lib/hooks";
 import { useFederatedOAuthStatus } from "@/lib/hooks/useFederatedOAuthStatus";
-import { FeedbackType } from "@/app/chat/interfaces";
 import { OnyxInitializingLoader } from "@/components/OnyxInitializingLoader";
 import { FeedbackModal } from "@/app/chat/components/modal/FeedbackModal";
 import { FiArrowDown } from "react-icons/fi";
@@ -50,16 +49,12 @@ import { useChatController } from "@/app/chat/hooks/useChatController";
 import { useAssistantController } from "@/app/chat/hooks/useAssistantController";
 import { useChatSessionController } from "@/app/chat/hooks/useChatSessionController";
 import { useDeepResearchToggle } from "@/app/chat/hooks/useDeepResearchToggle";
+import { useFeedbackController } from "@/app/chat/hooks/useFeedbackController";
 import {
   useChatSessionStore,
   useMaxTokens,
   useUncaughtError,
 } from "@/app/chat/stores/useChatSessionStore";
-import {
-  handleChatFeedback,
-  removeChatFeedback,
-} from "@/app/chat/services/lib";
-import { getMessageByMessageId } from "@/app/chat/services/messageTree";
 import {
   useCurrentChatState,
   useSubmittedMessage,
@@ -280,77 +275,8 @@ export function ChatPage({
   const filterManager = useFilters();
   const [isChatSearchModalOpen, setIsChatSearchModalOpen] = useState(false);
 
-  // Get the store action for updating message feedback
-  const updateCurrentMessageFeedback = useChatSessionStore(
-    (state) => state.updateCurrentMessageFeedback
-  );
-
-  // Optimistic feedback handler
-  const handleFeedbackChange = useCallback(
-    async (
-      messageId: number,
-      newFeedback: FeedbackType | null,
-      feedbackText?: string,
-      predefinedFeedback?: string
-    ) => {
-      // Get current feedback state for rollback on error
-      const { currentSessionId, sessions } = useChatSessionStore.getState();
-      const messageTree = currentSessionId
-        ? sessions.get(currentSessionId)?.messageTree
-        : undefined;
-      const previousFeedback = messageTree
-        ? getMessageByMessageId(messageTree, messageId)?.currentFeedback ?? null
-        : null;
-
-      // Optimistically update the UI
-      updateCurrentMessageFeedback(messageId, newFeedback);
-
-      try {
-        if (newFeedback === null) {
-          // Remove feedback
-          const response = await removeChatFeedback(messageId);
-          if (!response.ok) {
-            // Rollback on error
-            updateCurrentMessageFeedback(messageId, previousFeedback);
-            const errorData = await response.json();
-            setPopup({
-              message: `Failed to remove feedback - ${
-                errorData.detail || errorData.message
-              }`,
-              type: "error",
-            });
-          }
-        } else {
-          // Add/update feedback
-          const response = await handleChatFeedback(
-            messageId,
-            newFeedback,
-            feedbackText || "",
-            predefinedFeedback
-          );
-          if (!response.ok) {
-            // Rollback on error
-            updateCurrentMessageFeedback(messageId, previousFeedback);
-            const errorData = await response.json();
-            setPopup({
-              message: `Failed to submit feedback - ${
-                errorData.detail || errorData.message
-              }`,
-              type: "error",
-            });
-          }
-        }
-      } catch (error) {
-        // Rollback on network error
-        updateCurrentMessageFeedback(messageId, previousFeedback);
-        setPopup({
-          message: "Failed to submit feedback - network error",
-          type: "error",
-        });
-      }
-    },
-    [updateCurrentMessageFeedback, setPopup]
-  );
+  // Feedback controller with optimistic updates and error handling
+  const { handleFeedbackChange } = useFeedbackController({ setPopup });
 
   const [aboveHorizon, setAboveHorizon] = useState(false);
